Extract webdriver attributes methods to separate class
