+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-11-07 17:31:42 +0300 using RuboCop version 0.45.0.
+# on 2016-11-07 18:26:07 +0300 using RuboCop version 0.45.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,6 +12,7 @@ Lint/RescueException:
     - 'lib/onlyoffice_webdriver_wrapper/helpers/headless_helper.rb'
     - 'lib/onlyoffice_webdriver_wrapper/helpers/headless_helper/real_display_tools.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver.rb'
+    - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_attributes_helper.rb'
 
 # Offense count: 12
 Metrics/AbcSize:
@@ -20,7 +21,7 @@ Metrics/AbcSize:
 # Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 981
+  Max: 933
 
 # Offense count: 4
 Metrics/CyclomaticComplexity:
