+++ lib/onlyoffice_webdriver_wrapper/webdriver.rb
@@ -17,6 +17,7 @@ module OnlyofficeWebdriverWrapper
     include ChromeHelper
     include FirefoxHelper
     include RubyHelper
+    include WebdriverAttributesHelper
     include WebdriverHelper
     include WebdriverJsMethods
     include WebdriverUserAgentHelper
@@ -286,10 +287,6 @@ module OnlyofficeWebdriverWrapper
 
     def get_text_array(array_elements)
       get_elements(array_elements).map { |current_element| get_text(current_element) }
-    end
-
-    def get_element_by_parameter(elements, parameter_name, value)
-      elements.find { |current_element| value == get_attribute(current_element, parameter_name) }
     end
 
     def click(element)
@@ -873,13 +870,6 @@ module OnlyofficeWebdriverWrapper
       elements
     end
 
-    def get_index_of_elements_with_attribute(xpath, attribute, value, only_visible = true)
-      get_elements(xpath, only_visible).each_with_index do |element, index|
-        return (index + 1) if get_attribute(element, attribute).include?(value)
-      end
-      0
-    end
-
     def element_visible?(xpath_name)
       if xpath_name.is_a?(PageObject::Elements::Element) # PageObject always visible
         true
@@ -984,53 +974,6 @@ module OnlyofficeWebdriverWrapper
 
     def get_text_of_several_elements(xpath_several_elements)
       @driver.find_elements(:xpath, xpath_several_elements).map { |element| element.text unless element.text == '' }.compact
-    end
-
-    def attribute_exist?(xpath_name, attribute)
-      exist = false
-
-      element = xpath_name.is_a?(String) ? get_element(xpath_name) : xpath_name
-      if @browser == :ie
-        begin
-          element.attribute_value(attribute)
-          exist = true
-        rescue Exception
-          exist = false
-        end
-      else
-        begin
-          attribute_value = element.attribute(attribute)
-          exist = attribute_value.empty? || attribute_value.nil? ? false : true
-        rescue Exception
-          exist = false
-        end
-      end
-      exist
-    end
-
-    def get_attribute(xpath_name, attribute)
-      element = xpath_name.is_a?(Selenium::WebDriver::Element) ? xpath_name : get_element(xpath_name)
-
-      if element.nil?
-        webdriver_error("Webdriver.get_attribute(#{xpath_name}, #{attribute}) failed because element not found")
-      else
-        @browser == :ie ? element.attribute_value(attribute) : element.attribute(attribute)
-      end
-    end
-
-    def get_attribute_from_displayed_element(xpath_name, attribute)
-      @driver.find_elements(:xpath, xpath_name).each do |current_element|
-        return current_element.attribute(attribute) if current_element.displayed?
-      end
-      false
-    end
-
-    def get_attributes_of_several_elements(xpath_several_elements, attribute)
-      elements = @browser == :ie ? @driver.elements(:xpath, xpath_several_elements) : @driver.find_elements(:xpath, xpath_several_elements)
-
-      elements.map do |element|
-        @browser == :ie ? element.attribute_value(attribute) : element.attribute(attribute)
-      end
     end
 
     def get_style_parameter(xpath, parameter_name)
